# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Xihe API
  version: 1.0.0
  description: Canonical HTTP contract for Xihe public and service APIs.
servers:
  - url: /
paths:
...
  /api/v1/files/{fileId}:
    get: { operationId: getFile, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: fileId, required: true, schema: { type: string } }], responses: { '200': { description: File content } } }
    delete: { operationId: deleteFile, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: fileId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
  /api/v1/files/upload:
    post: { operationId: uploadFile, security: [{ bearerAuth: [] }], responses: { '201': { description: Uploaded } } }
  /api/v1/files/split-pdf:
    post: { operationId: splitPdf, security: [{ bearerAuth: [] }], responses: { '200': { description: Split result } } }
  /api/v1/rag/stats:
    get: { operationId: ragStats, security: [{ bearerAuth: [] }], responses: { '200': { description: RAG statistics } } }
  /api/v1/rag/ingest:
    post: { operationId: ragIngest, security: [{ bearerAuth: [] }], responses: { '200': { description: Ingest result } } }
  /api/v1/rag/search:
    post: { operationId: ragSearch, security: [{ bearerAuth: [] }], responses: { '200': { description: Search result } } }
  /api/v1/rag/documents/{docId}:
    delete: { operationId: ragDelete, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: docId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /internal/v1/agent/rag/ingest:
    post: { operationId: agentRagIngest, security: [{ bearerAuth: [] }], responses: { '200': { description: RAG result } } }
  /internal/v1/agent/rag/search:
    post: { operationId: agentRagSearch, security: [{ bearerAuth: [] }], responses: { '200': { description: RAG result } } }
  /internal/v1/agent/rag/stats:
    get: { operationId: agentRagStats, security: [{ bearerAuth: [] }], responses: { '200': { description: RAG statistics } } }
  /internal/v1/agent/rag/documents/{docId}:
    delete: { operationId: agentRagDelete, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: docId, required: true, schema: { type: string } }], responses: { '204': { description: Deleted } } }
...
  /internal/v1/runtime/workspaces:
    post: { operationId: registerRuntimeWorkspace, security: [{ bearerAuth: [] }], responses: { '200': { description: Registered } } }
...
  /internal/v1/runtime/workspaces/{workspaceId}/files/delete:
    post: { operationId: runtimeDeleteFile, security: [{ bearerAuth: [] }], parameters: [{ in: path, name: workspaceId, required: true, schema: { type: string } }], responses: { '200': { description: Deleted } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
...